ingestion/requirements.txt
requests
pandas
pyarrow
google-cloud-storage

dbt-core
dbt-bigquery
